Workspaced, created shared functions package
