+++ .nvmrc
@@ -1,1 +1,1 @@
-8.17.0
+12.14.1
