+++ functions/package.json
@@ -10,9 +10,6 @@
     "deploy": "firebase deploy --only functions",
     "logs": "firebase functions:log"
   },
-  "engines": {
-    "node": "8"
-  },
   "main": "lib/index.js",
   "dependencies": {
     "cors": "^2.8.5",
@@ -25,7 +22,8 @@
     "io-ts-reporters": "^1.2.2",
     "io-ts-types": "^0.5.11",
     "monocle-ts": "^2.3.3",
-    "newtype-ts": "^0.3.4"
+    "newtype-ts": "^0.3.4",
+    "solo-lib": "0.1.0"
   },
   "devDependencies": {
     "@types/cors": "^2.8.8",
@@ -34,7 +32,8 @@
     "@types/node": "12.0.0",
     "firebase-functions-test": "^0.2.2",
     "tslint": "^5.12.0",
-    "typescript": "^3.8.0"
+    "typescript": "3.9.7",
+    "prettier": "2.1.2"
   },
   "private": true
 }
