+++ functions/src/api/lobby/api/postCreateLobby.ts
@@ -1,7 +1,9 @@
 import { RequestHandler } from "express";
 import * as admin from "firebase-admin";
 import * as O from "fp-ts/Option";
-import { buildDeck, createAvatar, sortDeck } from "../../../helpers/game";
+import { buildDeck, sortDeck } from "solo-lib/lib/card";
+import { createAvatar } from "solo-lib/lib/player";
+
 import { ServerSession } from "../../../db/session";
 
 const codeGenerator = () => {
