+++ package.json
@@ -2,7 +2,7 @@
   "name": "minimal-solo",
   "version": "0.7.0",
   "private": true,
-  "workspaces": ["functions", "web"],
+  "workspaces": ["functions", "web", "solo"],
   "scripts": {
     "start:web": "yarn --cwd web start",
     "start:emulators": "firebase emulators:start"
