+++ solo/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "solo-lib",
+  "version": "0.1.0",
+  "description": "Minimal Solo's shared functions and utilities.",
+  "main": "lib/index.js",
+  "author": "mikelfcosta & davidpn11",
+  "license": "MIT",
+  "private": true,
+  "scripts": {
+    "build": "yarn tsc"
+  },
+  "devDependencies": {
+    "prettier": "2.1.2",
+    "typescript": "3.9.7"
+  },
+  "dependencies": {
+    "faker": "4.1.0",
+    "fp-ts": "2.8.4"
+  }
+}
