+++ web/package.json
@@ -16,6 +16,7 @@
     "react-scripts": "3.4.1",
     "redux": "^4.0.5",
     "redux-thunk": "^2.3.0",
+    "solo-lib": "0.1.0",
     "styled-components": "^5.1.0",
     "yup": "^0.28.5"
   },
@@ -68,10 +69,10 @@
     "@types/yup": "^0.28.1",
     "faker": "4.1.0",
     "polished": "^3.5.2",
-    "prettier": "^2.0.4",
+    "prettier": "2.1.2",
     "redux-devtools-extension": "^2.13.8",
     "storybook-dark-mode": "^0.4.1",
     "ts-node": "^8.10.1",
-    "typescript": "~3.7.2"
+    "typescript": "3.9.7"
   }
 }
